backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
psycopg==3.2.3
psycopg-binary==3.2.3
redis==5.2.1
sqlparse==0.5.3
typing_extensions==4.12.2
